Create todos service #1
Delete a Todo of a User
DELETE /users/{username}/todo/{todo_id}

+++ src/main/java/za/co/iherridge0/rest/webservices/todorestfulwebservices/todo/repository/ToDoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import za.co.iherridge0.rest.webservices.todorestfulwebservices.todo.entity.Todo;
 
-public interface ToDoRepository extends JpaRepository<Todo, Integer> {
+public interface ToDoRepository extends JpaRepository<Todo, Long> {
 	
 }
